+++ modules/tribes/client/services/tribes.client.service.js
@@ -1,0 +1,20 @@
+(function () {
+  'use strict';
+
+  angular
+    .module('tribes')
+    .factory('TribesService', TribesService);
+
+  /* @ngInject */
+  function TribesService($resource) {
+    return $resource('/api/tribes', {
+      limit: 50
+    }, {
+      'query': {
+        method: 'GET',
+        isArray: true
+      }
+    });
+  }
+
+}());