+++ modules/tribes/server/routes/tribes.server.routes.js
@@ -1,0 +1,19 @@
+'use strict';
+
+/**
+ * Module dependencies.
+ */
+var tribesPolicy = require('../policies/tribes.server.policy'),
+    tribes = require('../controllers/tribes.server.controller');
+
+module.exports = function (app) {
+
+  app.route('/api/tribes').all(tribesPolicy.isAllowed)
+    .get(tribes.listTribes);
+
+  app.route('/api/tribes/:tribe').all(tribesPolicy.isAllowed)
+    .get(tribes.getTribe);
+
+  // Finish by binding the tribes middleware
+  app.param('tribe', tribes.tribeBySlug);
+};