+++ modules/users/tests/server/user.server.routes.tests.js
@@ -4,9 +4,8 @@ var should = require('should'),
     request = require('supertest'),
     path = require('path'),
     mongoose = require('mongoose'),
-    semver = require('semver'),
     User = mongoose.model('User'),
-    Tag = mongoose.model('Tag'),
+    Tribe = mongoose.model('Tribe'),
     inviteCodeService = require(path.resolve('./modules/users/server/services/invite-codes.server.service')),
     config = require(path.resolve('./config/config')),
     express = require(path.resolve('./config/lib/express')),
@@ -176,7 +175,7 @@ describe('User CRUD tests', function () {
                 return done(confirmEmailPostErr);
               }
 
-              confirmEmailGetRes.text.should.equal(redirectMessage('/confirm-email/' + userRes1.emailToken));
+              confirmEmailGetRes.text.should.equal('Found. Redirecting to /confirm-email/' + userRes1.emailToken);
 
               // POST does the actual job
               agent.post('/api/auth/confirm-email/' + userRes1.emailToken)
@@ -243,7 +242,7 @@ describe('User CRUD tests', function () {
                 return done(confirmEmailPostErr);
               }
 
-              confirmEmailGetRes.text.should.equal(redirectMessage('/confirm-email-invalid'));
+              confirmEmailGetRes.text.should.equal('Found. Redirecting to /confirm-email-invalid');
 
               // POST does the actual job
               agent.post('/api/auth/confirm-email/WRONG_TOKEN')
@@ -286,7 +285,7 @@ describe('User CRUD tests', function () {
             }
 
             signoutRes.redirect.should.equal(true);
-            signoutRes.text.should.equal(redirectMessage('/'));
+            signoutRes.text.should.equal('Found. Redirecting to /');
 
             return done();
           });
@@ -317,7 +316,7 @@ describe('User CRUD tests', function () {
             }
 
             signoutRes.redirect.should.equal(true);
-            signoutRes.text.should.equal(redirectMessage('/'));
+            signoutRes.text.should.equal('Found. Redirecting to /');
 
             return done();
           });
@@ -1365,7 +1364,7 @@ describe('User CRUD tests', function () {
       });
   });
 
-  it('should be able to join a tribe with "is" relation', function (done) {
+  it('should be able to join a tribe', function (done) {
     agent.post('/api/auth/signin')
       .send(credentials)
       .expect(200)
@@ -1376,44 +1375,39 @@ describe('User CRUD tests', function () {
         }
 
         // Create test tribe
-        var tribe = new Tag({
-          label: 'Awesome Tribe',
-          tribe: true
+        var tribe = new Tribe({
+          label: 'Awesome Tribe'
         });
 
         // Add tribe to test DB
         tribe.save(function (err, tribe) {
           should.not.exist(err);
 
-          agent.post('/api/users/memberships')
-            .send({
-              id: tribe._id,
-              relation: 'is'
-            })
+          agent.post('/api/users/memberships/' + tribe._id)
+            .send()
             .expect(200)
-            .end(function (userTagErr, userTagRes) {
+            .end(function (userTribeErr, userTribeRes) {
 
               // Handle joining tag error
-              if (userTagErr) {
-                return done(userTagErr);
+              if (userTribeErr) {
+                return done(userTribeErr);
               }
 
               // Confirmation message
-              userTagRes.body.message.should.be.equal('Joined tribe.');
+              userTribeRes.body.message.should.be.equal('Joined tribe.');
 
               // It should return correct tribe with new count
-              userTagRes.body.tag._id.should.be.equal(tribe._id.toString());
-              userTagRes.body.tag.count.should.be.equal(1);
-              userTagRes.body.tag.label.should.be.equal('Awesome Tribe');
-              userTagRes.body.tag.slug.should.be.equal('awesome-tribe');
-              should.exist(userTagRes.body.tag.color);
+              userTribeRes.body.tribe._id.should.be.equal(tribe._id.toString());
+              userTribeRes.body.tribe.count.should.be.equal(1);
+              userTribeRes.body.tribe.label.should.be.equal('Awesome Tribe');
+              userTribeRes.body.tribe.slug.should.be.equal('awesome-tribe');
+              should.exist(userTribeRes.body.tribe.color);
 
               // It should return updated user
-              userTagRes.body.user.username.should.be.equal(credentials.username.toLowerCase());
-              userTagRes.body.user.memberIds[0].should.be.equal(tribe._id.toString());
-              userTagRes.body.user.member[0].tag.should.be.equal(tribe._id.toString());
-              userTagRes.body.user.member[0].relation.should.be.equal('is');
-              should.exist(userTagRes.body.user.member[0].since);
+              userTribeRes.body.user.username.should.be.equal(credentials.username.toLowerCase());
+              userTribeRes.body.user.memberIds[0].should.be.equal(tribe._id.toString());
+              userTribeRes.body.user.member[0].tribe.should.be.equal(tribe._id.toString());
+              should.exist(userTribeRes.body.user.member[0].since);
 
               return done();
             });
@@ -1421,7 +1415,7 @@ describe('User CRUD tests', function () {
       });
   });
 
-  it('should be able to join a tribe with "likes" relation', function (done) {
+  it('should be able to leave tribes', function (done) {
     agent.post('/api/auth/signin')
       .send(credentials)
       .expect(200)
@@ -1432,96 +1426,35 @@ describe('User CRUD tests', function () {
         }
 
         // Create test tribe
-        var tribe = new Tag({
-          label: 'Awesome Tribe',
-          tribe: true
+        var tribe = new Tribe({
+          label: 'Hitchhikers'
         });
 
         // Add tribe to test DB
         tribe.save(function (err, tribe) {
           should.not.exist(err);
 
-          agent.post('/api/users/memberships')
-            .send({
-              id: tribe._id,
-              relation: 'likes'
-            })
+          // Join tribe
+          agent.post('/api/users/memberships/' + tribe._id)
+            .send()
             .expect(200)
-            .end(function (userTagErr, userTagRes) {
+            .end(function (userTribeJoinErr, userTribeJoinRes) {
 
               // Handle joining tag error
-              if (userTagErr) {
-                return done(userTagErr);
-              }
-
-              // Confirmation message
-              userTagRes.body.message.should.be.equal('Joined tribe.');
-
-              // It should return correct tribe with new count
-              userTagRes.body.tag._id.should.be.equal(tribe._id.toString());
-              userTagRes.body.tag.count.should.be.equal(1);
-              userTagRes.body.tag.label.should.be.equal('Awesome Tribe');
-              userTagRes.body.tag.slug.should.be.equal('awesome-tribe');
-              should.exist(userTagRes.body.tag.color);
-
-              // It should return updated user
-              userTagRes.body.user.username.should.be.equal(credentials.username.toLowerCase());
-              userTagRes.body.user.memberIds[0].should.be.equal(tribe._id.toString());
-              userTagRes.body.user.member[0].tag.should.be.equal(tribe._id.toString());
-              userTagRes.body.user.member[0].relation.should.be.equal('likes');
-              should.exist(userTagRes.body.user.member[0].since);
-
-              return done();
-            });
-        });
-      });
-  });
-
-  it('should be able to leave tribe', function (done) {
-    agent.post('/api/auth/signin')
-      .send(credentials)
-      .expect(200)
-      .end(function (signinErr) {
-        // Handle signin error
-        if (signinErr) {
-          return done(signinErr);
-        }
-
-        // Create test tribe
-        var tribe = new Tag({
-          label: 'Hitchhikers',
-          tribe: true
-        });
-
-        // Add tribe to test DB
-        tribe.save(function (err, tribe) {
-          should.not.exist(err);
-
-          agent.post('/api/users/memberships')
-            .send({
-              id: tribe._id,
-              relation: 'is'
-            })
-            .expect(200)
-            .end(function (userTagJoinErr, userTagJoinRes) {
-
-              // Handle joining tag error
-              if (userTagJoinErr) {
-                return done(userTagJoinErr);
+              if (userTribeJoinErr) {
+                return done(userTribeJoinErr);
               }
 
               // Count +1
-              userTagJoinRes.body.tag.count.should.be.equal(1);
+              userTribeJoinRes.body.tribe.count.should.be.equal(1);
 
               // User is now member of tribe
-              userTagJoinRes.body.user.memberIds.length.should.be.equal(1);
-              userTagJoinRes.body.user.member.length.should.be.equal(1);
+              userTribeJoinRes.body.user.memberIds.length.should.be.equal(1);
+              userTribeJoinRes.body.user.member.length.should.be.equal(1);
 
-              agent.post('/api/users/memberships')
-                .send({
-                  id: tribe._id,
-                  relation: 'leave'
-                })
+              // Leave tribe
+              agent.delete('/api/users/memberships/' + tribe._id)
+                .send()
                 .expect(200)
                 .end(function (userTagLeaveErr, userTagLeaveRes) {
 
@@ -1531,9 +1464,9 @@ describe('User CRUD tests', function () {
                   }
 
                   // Count -1
-                  userTagLeaveRes.body.tag.count.should.be.equal(0);
+                  userTagLeaveRes.body.tribe.count.should.be.equal(0);
 
-                  // No more tags left on user's array
+                  // No more tribes left on user's array
                   userTagLeaveRes.body.user.memberIds.length.should.be.equal(0);
                   userTagLeaveRes.body.user.member.length.should.be.equal(0);
 
@@ -1556,40 +1489,35 @@ describe('User CRUD tests', function () {
         }
 
         // Create test tribe
-        var tribe = new Tag({
-          label: 'Russian literature students',
-          tribe: true
+        var tribe = new Tribe({
+          label: 'Russian literature students'
         });
 
         // Add tribe to test DB
         tribe.save(function (err, tribe) {
           should.not.exist(err);
 
-          agent.post('/api/users/memberships')
-            .send({
-              id: tribe._id,
-              relation: 'is'
-            })
+          // Join tribe
+          agent.post('/api/users/memberships/' + tribe._id)
+            .send()
             .expect(200)
-            .end(function (userTagJoinErr, userTagJoinRes) {
+            .end(function (userTribeJoinErr, userTribeJoinRes) {
 
               // Handle joining tag error
-              if (userTagJoinErr) {
-                return done(userTagJoinErr);
+              if (userTribeJoinErr) {
+                return done(userTribeJoinErr);
               }
 
               // Count +1
-              userTagJoinRes.body.tag.count.should.be.equal(1);
+              userTribeJoinRes.body.tribe.count.should.be.equal(1);
 
               // User is now member of tribe
-              userTagJoinRes.body.user.memberIds.length.should.be.equal(1);
-              userTagJoinRes.body.user.member.length.should.be.equal(1);
+              userTribeJoinRes.body.user.memberIds.length.should.be.equal(1);
+              userTribeJoinRes.body.user.member.length.should.be.equal(1);
 
-              agent.post('/api/users/memberships')
-                .send({
-                  id: tribe._id,
-                  relation: 'is'
-                })
+              // Join tribe again
+              agent.post('/api/users/memberships/' + tribe._id)
+                .send()
                 .expect(409)
                 .end(function (userTagJoin2Err, userTagJoin2Res) {
 
@@ -1598,7 +1526,7 @@ describe('User CRUD tests', function () {
                     return done(userTagJoin2Err);
                   }
 
-                  userTagJoin2Res.body.message.should.be.equal('Conflict.');
+                  userTagJoin2Res.body.message.should.be.equal('You are already a member of this tribe.');
 
                   return done();
                 });
@@ -1608,7 +1536,7 @@ describe('User CRUD tests', function () {
       });
   });
 
-  it('should be able to show error if trying to leave tribe user is not member of', function (done) {
+  it('should be able to show error if trying to leave tribe user is not member', function (done) {
     agent.post('/api/auth/signin')
       .send(credentials)
       .expect(200)
@@ -1619,70 +1547,26 @@ describe('User CRUD tests', function () {
         }
 
         // Create test tribe
-        var tribe = new Tag({
-          label: 'Japanese linguistics',
-          tribe: true
+        var tribe = new Tribe({
+          label: 'Japanese linguistics'
         });
 
         // Add tribe to test DB
         tribe.save(function (err, tribe) {
           should.not.exist(err);
 
-          agent.post('/api/users/memberships')
-            .send({
-              id: tribe._id,
-              relation: 'leave'
-            })
+          // Leave tribe
+          agent.delete('/api/users/memberships/' + tribe._id)
+            .send()
             .expect(409)
-            .end(function (userTagJoinErr, userTagJoinRes) {
+            .end(function (userTribeJoinErr, userTribeJoinRes) {
 
               // Handle joining tag error
-              if (userTagJoinErr) {
-                return done(userTagJoinErr);
+              if (userTribeJoinErr) {
+                return done(userTribeJoinErr);
               }
 
-              userTagJoinRes.body.message.should.be.equal('Conflict.');
-
-              return done();
-            });
-        });
-      });
-  });
-
-  it('should be able to show error if sending an action with odd "relation" info', function (done) {
-    agent.post('/api/auth/signin')
-      .send(credentials)
-      .expect(200)
-      .end(function (signinErr) {
-        // Handle signin error
-        if (signinErr) {
-          return done(signinErr);
-        }
-
-        // Create test tribe
-        var tribe = new Tag({
-          label: 'Space Squirrels',
-          tribe: true
-        });
-
-        // Add tribe to test DB
-        tribe.save(function (err, tribe) {
-          should.not.exist(err);
-
-          agent.post('/api/users/memberships')
-            .send({
-              id: tribe._id,
-              relation: 'wtf'
-            })
-            .expect(400)
-            .end(function (userTagJoinErr, userTagJoinRes) {
-
-              // Handle joining tag error
-              if (userTagJoinErr) {
-                return done(userTagJoinErr);
-              }
-
-              userTagJoinRes.body.message.should.be.equal('Missing relation info.');
+              userTribeJoinRes.body.message.should.be.equal('You are not a member of this tribe.');
 
               return done();
             });
@@ -1700,20 +1584,18 @@ describe('User CRUD tests', function () {
           return done(signinErr);
         }
 
-        agent.post('/api/users/memberships')
-          .send({
-            id: '572a3d36f905fe5c53bf1d1f',
-            relation: 'is'
-          })
+        // Join tribe
+        agent.post('/api/users/memberships/572a3d36f905fe5c53bf1d1f')
+          .send()
           .expect(400)
-          .end(function (userTagJoinErr, userTagJoinRes) {
+          .end(function (userTribeJoinErr, userTribeJoinRes) {
 
             // Handle joining tag error
-            if (userTagJoinErr) {
-              return done(userTagJoinErr);
+            if (userTribeJoinErr) {
+              return done(userTribeJoinErr);
             }
 
-            userTagJoinRes.body.message.should.be.equal('Bad request.');
+            userTribeJoinRes.body.message.should.be.equal('Bad request.');
 
             return done();
           });
@@ -1730,20 +1612,18 @@ describe('User CRUD tests', function () {
           return done(signinErr);
         }
 
-        agent.post('/api/users/memberships')
-          .send({
-            id: '123456',
-            relation: 'is'
-          })
+        // Join tribe
+        agent.post('/api/users/memberships/123456')
+          .send()
           .expect(400)
-          .end(function (userTagJoinErr, userTagJoinRes) {
+          .end(function (userTribeJoinErr, userTribeJoinRes) {
 
             // Handle joining tag error
-            if (userTagJoinErr) {
-              return done(userTagJoinErr);
+            if (userTribeJoinErr) {
+              return done(userTribeJoinErr);
             }
 
-            userTagJoinRes.body.message.should.be.equal('Cannot interpret id.');
+            userTribeJoinRes.body.message.should.be.equal('Cannot interpret id.');
 
             return done();
           });
@@ -1881,20 +1761,7 @@ describe('User CRUD tests', function () {
 
   afterEach(function (done) {
     User.remove().exec(function () {
-      Tag.remove().exec(done);
+      Tribe.remove().exec(done);
     });
   });
 });
-
-/**
- * Returns the NodeJS redirect text for any version.
- */
-function redirectMessage(url) {
-  // NodeJS v4 changed the status code representation so we must check
-  // before asserting, to be comptabile with all node versions.
-  if (semver.satisfies(process.versions.node, '>=4.0.0')) {
-    return 'Found. Redirecting to ' + url;
-  } else {
-    return 'Moved Temporarily. Redirecting to ' + url;
-  }
-}